refactor: rename _base -> _vis_model (#26)
* refactor: rename _base -> _vis_model

* fix: fix one more
